Update harvested area data handling.
Add function to download MIRCA2000 harvested_area_grids data and unzip first level zip and second level .gz files. Then use the file to create raster stack.
The reason to switch to download data from source is that the data is large, and raster::stack does not load the data to object. So users will encounter error without the actual harvested are data.